Map stylizer script.
This (pretty hacky) script reads in a map saved in DXF format with some
specific conventions, and writes out an SVG file with a stylized rendering
of the map.

The idea here is to do the layout bits in a CAD package, where there are
more useful tools for accurately placing objects, creating nice curves,
etc. The result of that can then be systematically transformed into a
nicer rendering, to let the computer do the fiddly presentation work
and let me concentrate on the fundamental geometry.